+++ ui/src/app/core/accessories/types/matter/robotic-vacuum-cleaner/robotic-vacuum-cleaner.component.ts
@@ -5,7 +5,8 @@ import { TranslatePipe } from '@ngx-translate/core'
 import { ServiceTypeX } from '@/app/core/accessories/accessories.interfaces'
 import { AccessoriesService } from '@/app/core/accessories/accessories.service'
 import { ACCESSORY_MANAGE_MODAL_DATA } from '@/app/core/accessories/types/base-manage.component'
-import { controlDevice, getDeviceActiveState, getDeviceStatusText, isOnOffDevice } from '@/app/core/accessories/types/matter/matter-device.utils'
+import { RvcOperationalState } from '@/app/core/accessories/types/matter/matter-device.constants'
+import { controlDevice, getDeviceActiveState, getDeviceStatusText, getRvcOperationalState, isOnOffDevice } from '@/app/core/accessories/types/matter/matter-device.utils'
 import { RoboticVacuumCleanerManageComponent } from '@/app/core/accessories/types/matter/robotic-vacuum-cleaner/robotic-vacuum-cleaner.manage.component'
 import { LongClickDirective } from '@/app/core/directives/long-click.directive'
 
@@ -61,6 +62,10 @@ export class RoboticVacuumCleanerComponent {
   }
 
   public readonly isActive = computed(() => getDeviceActiveState(this.service()))
+
+  public readonly isRunning = computed(() => getRvcOperationalState(this.service()) === RvcOperationalState.Running)
+
+  public readonly isCharging = computed(() => getRvcOperationalState(this.service()) === RvcOperationalState.Charging)
 
   public readonly statusText = computed(() => getDeviceStatusText(this.service()))